requirements.txt
Brotli==1.0.9
colorama==0.4.5
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
itsdangerous==2.1.2
MarkupSafe==2.1.1
numpy==1.23.0
plotly==5.9.0
pytz==2022.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.1.2
zipp==3.8.0
